refactor to use popupview toolkit component
