+++ app/build.gradle
@@ -42,6 +42,7 @@ dependencies {
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.constraintlayout:constraintlayout:2.0.4'
     implementation 'com.esri.arcgisruntime:arcgis-android:100.10.0'
+    implementation 'com.esri.arcgisruntime:arcgis-android-toolkit:100.11.0-3131'
     implementation 'androidx.lifecycle:lifecycle-extensions:2.2.0'
     implementation 'androidx.navigation:navigation-fragment-ktx:2.3.3'
     implementation 'androidx.navigation:navigation-ui-ktx:2.3.3'
